move tests to top-level directory
