+++ numth/modular/__init__.py
@@ -1,20 +1,20 @@
 #   numth/modular/__init__.py
 #===========================================================
 
-from .sqrt import (
-    mod_sqrt,
-    mod_sqrt_minus_one_wilson,
-    mod_sqrt_minus_one_legendre,
-    mod_sqrt_when_three_mod_four,
-    mod_sqrt_tonelli_shanks,
-    mod_sqrt_cipolla
+from .carmichael_lambda import (
+    carmichael_lambda
 )
 
 from .euler_phi import (
     euler_phi
 )
 
-from .carmichael_lambda import (
-    carmichael_lambda
+from .sqrt import (
+    mod_sqrt,
+    mod_sqrt_minus_one_wilson,
+    mod_sqrt_minus_one_legendre,
+    mod_sqrt_when_three_mod_four,
+    mod_sqrt_cipolla
+    mod_sqrt_tonelli_shanks,
 )
