--- numth/tests/__init__.py
@@ -1,2 +1,0 @@
-#   numth/tests/__init__.py
-
