+++ tests/env.py
@@ -1,0 +1,14 @@
+#   env.py
+#===========================================================
+import sys
+import os
+#===========================================================
+
+sys.path.append(
+    os.path.dirname(
+        os.path.dirname(
+            os.path.abspath(__file__)
+        )
+    )
+)
+
